Move integration test helpers to test module
